[hotfix] Rename org.apache.flink.yarn.util.YarnTestUtils into o.a.f.y.u.TestUtils
Renamed o.a.f.y.u.YarnTestUtils into o.a.f.y.u.TestUtils in order to avoid a naming conflict with
o.a.f.y.YarnTestUtils from the flink-yarn module.
